refactor(x509-claims): inline the policy module into the crate root
The root was 28 lines that mostly re-exported the module below it, and the
split forced nine items to be `pub` just so the root could name them. With
one file, the internal helpers go back to being private.
